requirements.txt
torch
torchvision
opencv-python
numpy
pillow
timm
scipy
pyyaml
scikit-image
transformers